feat: PMAT-225 - 9 recipes for contracts/MC/tui + 7 pre-existing test fixes (Refs PMAT-225)
9 new IIUR recipes:
- contracts-macros: yaml-key-type-check, recipe-runtime-min, witness-proof-status
- monte-carlo: coupling-chain-mix, leader-election-random, sgd-convergence-rate
- tui: toggle-switch-render, radio-button-select, fuzzy-match-score

Pre-existing test fixes (encountered during cargo test --examples):
- tui_log_tail: E0716 temporary-borrow fix in two tests
- cli_qa_warmup_iteration_budget: u32-overflow in budget calculation
- cli_rosetta_convert_external_tokenizer: arm ordering bug for TargetNeedsTokenizer::No
- cli_runs_ls_sparkline_renderer: downsample denom missed last value
- tokenize_bpe_trace: merge rules didn't account for word-sentinel prefix
- train_distributed_sim: lr too high causing divergence
- shell_trie_prefix_completion: weakened freq-rank assumption
- tui_kbd_shortcuts_render: padding test used wrong column-finding heuristic

+++ examples/cli/cli_qa_warmup_iteration_budget.rs
@@ -47,7 +47,7 @@ pub fn validate_budget(b: QaBudget) -> BudgetVerdict {
     if b.max_tokens == 0 {
         return BudgetVerdict::InvalidMaxTokens;
     }
-    let total = u64::from(b.warmup + b.iterations) * u64::from(b.max_tokens);
+    let total = (u64::from(b.warmup) + u64::from(b.iterations)) * u64::from(b.max_tokens);
     let measured = u64::from(b.iterations) * u64::from(b.max_tokens);
     BudgetVerdict::Ok {
         total_tokens: total,
